+++ src/cf.data.pre
@@ -3549,7 +3549,7 @@ COMMENT_END
 NAME: cache_peer
 TYPE: peer
 DEFAULT: none
-LOC: Config.peers
+LOC: Config.cachePeers
 DOC_START
 	To specify other caches in a hierarchy, use the format:
