+++ src/icmp/net_db.cc
@@ -1299,7 +1299,8 @@ netdbExchangeStart(void *data)
 static CachePeer *
 findUsableParentAtHostname(PeerSelector *ps, const char * const hostname, const HttpRequest &request)
 {
-    for (auto p = Config.peers; p; p = p->next) {
+    for (const auto &peer: Config.cachePeers) {
+        auto p = peer.get();
         // Both fields should be lowercase, but no code ensures that invariant.
         // TODO: net_db_peer should point to CachePeer instead of its hostname!
         if (strcasecmp(p->host, hostname) != 0)
