+++ src/neighbors.cc
@@ -97,13 +97,12 @@ whichPeer(const Ip::Address &from)
 {
     int j;
 
-    CachePeer *p = nullptr;
     debugs(15, 3, "whichPeer: from " << from);
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         for (j = 0; j < p->n_addresses; ++j) {
             if (from == p->addresses[j] && from.port() == p->icp.port) {
-                return p;
+                return p.get();
             }
         }
     }
@@ -272,11 +271,10 @@ peerHTTPOkay(const CachePeer * p, PeerSelector * ps)
 int
 neighborsCount(PeerSelector *ps)
 {
-    CachePeer *p = nullptr;
     int count = 0;
 
-    for (p = Config.peers; p; p = p->next)
-        if (peerWouldBePinged(p, ps))
+    for (const auto &p: Config.cachePeers)
+        if (peerWouldBePinged(p.get(), ps))
             ++count;
 
     debugs(15, 3, "neighborsCount: " << count);
@@ -290,9 +288,9 @@ getFirstUpParent(PeerSelector *ps)
     assert(ps);
     HttpRequest *request = ps->request;
 
-    CachePeer *p = nullptr;
+    for (const auto &peer: Config.cachePeers) {
+        auto p = peer.get();
 
-    for (p = Config.peers; p; p = p->next) {
         if (!neighborUp(p))
             continue;
 
@@ -302,11 +300,11 @@ getFirstUpParent(PeerSelector *ps)
         if (!peerHTTPOkay(p, ps))
             continue;
 
-        break;
+        debugs(15, 3, "returning " << RawPointer(p).orNil());
+        return p;
     }
 
-    debugs(15, 3, "returning " << RawPointer(p).orNil());
-    return p;
+    return nullptr;
 }
 
 CachePeer *
@@ -315,10 +313,10 @@ getRoundRobinParent(PeerSelector *ps)
     assert(ps);
     HttpRequest *request = ps->request;
 
-    CachePeer *p;
     CachePeer *q = nullptr;
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &peer: Config.cachePeers) {
+        auto p = peer.get();
         if (!p->options.roundrobin)
             continue;
 
@@ -357,11 +355,12 @@ getWeightedRoundRobinParent(PeerSelector *ps)
     assert(ps);
     HttpRequest *request = ps->request;
 
-    CachePeer *p;
     CachePeer *q = nullptr;
     int weighted_rtt;
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &peer: Config.cachePeers) {
+        auto p = peer.get();
+
         if (!p->options.weighted_roundrobin)
             continue;
 
@@ -378,11 +377,11 @@ getWeightedRoundRobinParent(PeerSelector *ps)
     }
 
     if (q && q->rr_count > 1000000)
-        for (p = Config.peers; p; p = p->next) {
+        for (const auto &p: Config.cachePeers) {
             if (!p->options.weighted_roundrobin)
                 continue;
 
-            if (neighborType(p, request->url) != PEER_PARENT)
+            if (neighborType(p.get(), request->url) != PEER_PARENT)
                 continue;
 
             p->rr_count = 0;
@@ -446,10 +445,8 @@ peerClearRRStart(void)
 void
 peerClearRR()
 {
-    CachePeer *p = nullptr;
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers)
         p->rr_count = 1;
-    }
 }
 
 void
@@ -477,9 +474,9 @@ getDefaultParent(PeerSelector *ps)
     assert(ps);
     HttpRequest *request = ps->request;
 
-    CachePeer *p = nullptr;
+    for (const auto &peer: Config.cachePeers) {
+        const auto p = peer.get();
 
-    for (p = Config.peers; p; p = p->next) {
         if (neighborType(p, request->url) != PEER_PARENT)
             continue;
 
@@ -497,18 +494,6 @@ getDefaultParent(PeerSelector *ps)
     // TODO: Refactor similar get*() functions to use our return/reporting style
     debugs(15, 3, "none found");
     return nullptr;
-}
-
-CachePeer *
-getNextPeer(CachePeer * p)
-{
-    return p->next;
-}
-
-CachePeer *
-getFirstPeer(void)
-{
-    return Config.peers;
 }
 
 static void
@@ -551,16 +536,12 @@ neighbors_init(void)
 {
     struct servent *sep = nullptr;
     const char *me = getMyHostname();
-    CachePeer *thisPeer = nullptr;
-    CachePeer *next = nullptr;
 
     neighborsRegisterWithCacheManager();
 
     if (Comm::IsConnOpen(icpIncomingConn)) {
 
-        for (thisPeer = Config.peers; thisPeer; thisPeer = next) {
-            next = thisPeer->next;
-
+        for (const auto &thisPeer: Config.cachePeers) {
             if (0 != strcmp(thisPeer->host, me))
                 continue;
 
@@ -571,7 +552,7 @@ neighbors_init(void)
                 debugs(15, DBG_IMPORTANT, "WARNING: Peer looks like this host." <<
                        Debug::Extra << "Ignoring cache_peer " << *thisPeer);
 
-                neighborRemove(thisPeer);
+                neighborRemove(thisPeer.get());
             }
         }
     }
@@ -1098,14 +1079,11 @@ neighborsUdpAck(const cache_key * key, icp_common_t * header, const Ip::Address 
 CachePeer *
 findCachePeerByName(const char * const name)
 {
-    CachePeer *p = nullptr;
-
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         if (!strcasecmp(name, p->name))
-            break;
+            return p.get();
     }
-
-    return p;
+    return nullptr;
 }
 
 int
@@ -1209,8 +1187,6 @@ peerDNSConfigure(const ipcache_addrs *ia, const Dns::LookupDetails &, void *data
 static void
 peerRefreshDNS(void *data)
 {
-    CachePeer *p = nullptr;
-
     if (eventFind(peerRefreshDNS, nullptr))
         eventDelete(peerRefreshDNS, nullptr);
 
@@ -1220,8 +1196,8 @@ peerRefreshDNS(void *data)
         return;
     }
 
-    for (p = Config.peers; p; p = p->next)
-        ipcache_nbgethostbyname(p->host, peerDNSConfigure, p);
+    for (const auto &p: Config.cachePeers)
+        ipcache_nbgethostbyname(p->host, peerDNSConfigure, p.get());
 
     /* Reconfigure the peers every hour */
     eventAddIsh("peerRefreshDNS", peerRefreshDNS, nullptr, 3600.0, 1);
@@ -1720,10 +1696,9 @@ neighborsHtcpReply(const cache_key * key, HtcpReplyData * htcp, const Ip::Addres
 void
 neighborsHtcpClear(StoreEntry * e, HttpRequest * req, const HttpRequestMethod &method, htcp_clr_reason reason)
 {
-    CachePeer *p;
     char buf[128];
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         if (!p->options.htcp) {
             continue;
         }
@@ -1734,7 +1709,7 @@ neighborsHtcpClear(StoreEntry * e, HttpRequest * req, const HttpRequestMethod &m
             continue;
         }
         debugs(15, 3, "neighborsHtcpClear: sending CLR to " << p->in_addr.toUrl(buf, 128));
-        htcpClear(e, req, method, p, reason);
+        htcpClear(e, req, method, p.get(), reason);
     }
 }
