+++ src/peer_userhash.cc
@@ -47,8 +47,6 @@ peerUserHashInit(void)
     int K;
     int k;
     double P_last, X_last, Xn;
-    CachePeer *p;
-    CachePeer **P;
     char *t;
     /* Clean up */
 
@@ -62,7 +60,7 @@ peerUserHashInit(void)
 
     peerUserHashRegisterWithCacheManager();
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         if (!p->options.userhash)
             continue;
 
@@ -81,8 +79,9 @@ peerUserHashInit(void)
 
     userhash_peers = (CachePeer **)xcalloc(n_userhash_peers, sizeof(*userhash_peers));
 
+    auto P = userhash_peers;
     /* Build a list of the found peers and calculate hashes and load factors */
-    for (P = userhash_peers, p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         if (!p->options.userhash)
             continue;
 
@@ -106,7 +105,7 @@ peerUserHashInit(void)
             p->userhash.load_factor = 0.0;
 
         /* add it to our list of peers */
-        *P++ = cbdataReference(p);
+        *P++ = cbdataReference(p.get());
     }
 
     /* Sort our list on weight */
@@ -130,7 +129,7 @@ peerUserHashInit(void)
 
     for (k = 1; k <= K; ++k) {
         double Kk1 = (double) (K - k + 1);
-        p = userhash_peers[k - 1];
+        auto p = userhash_peers[k - 1];
         p->userhash.load_multiplier = (Kk1 * (p->userhash.load_factor - P_last)) / Xn;
         p->userhash.load_multiplier += pow(X_last, Kk1);
         p->userhash.load_multiplier = pow(p->userhash.load_multiplier, 1.0 / Kk1);
@@ -203,7 +202,6 @@ peerUserHashSelectParent(PeerSelector *ps)
 static void
 peerUserHashCachemgr(StoreEntry * sentry)
 {
-    CachePeer *p;
     int sumfetches = 0;
     storeAppendPrintf(sentry, "%24s %10s %10s %10s %10s\n",
                       "Hostname",
@@ -212,10 +210,10 @@ peerUserHashCachemgr(StoreEntry * sentry)
                       "Factor",
                       "Actual");
 
-    for (p = Config.peers; p; p = p->next)
+    for (const auto &p: Config.cachePeers)
         sumfetches += p->stats.fetches;
 
-    for (p = Config.peers; p; p = p->next) {
+    for (const auto &p: Config.cachePeers) {
         storeAppendPrintf(sentry, "%24s %10x %10f %10f %10f\n",
                           p->name, p->userhash.hash,
                           p->userhash.load_multiplier,
