+++ src/snmp_agent.cc
@@ -190,14 +190,14 @@ snmp_meshPtblFn(variable_list * Var, snint * ErrP)
     Ip::Address laddr;
     char *cp = nullptr;
     CachePeer *p = nullptr;
-    int cnt = 0;
     debugs(49, 5, "snmp_meshPtblFn: peer " << Var->name[LEN_SQ_MESH + 3] << " requested!");
     *ErrP = SNMP_ERR_NOERROR;
 
     u_int index = Var->name[LEN_SQ_MESH + 3] ;
-    for (p = Config.peers; p != nullptr; p = p->next, ++cnt) {
+    for (const auto &peer: Config.cachePeers) {
         if (p->index == index) {
-            laddr = p->in_addr ;
+            laddr = peer->in_addr ;
+            p = peer.get();
             break;
         }
     }
